Invalidate cache when a class's superclass is reinitialized
I was under the impression that this would cause subclasses to be reinitialized,
but this is not the case. They will have their inheritance finalized again,
but (a) there's no way to depend (with dependent maintenance) on a class having
its inheritance re-finalized, and (b) there's no guarantee that the actual
function FINALIZE-INHERITANCE is called to do this.

So I came up with a terrible hack.